Fixes issue #8
The root cause of that bug was that the transformation code wasn't even
attempting to reuse RDG nodes already created for the fragment
references in the model.

Hence, I changed the method `SequenceDiagramTransformer.transformSD()`
to take only a SequenceDiagram as input and to infer the RDG node ID
from it. This way, the behavior became consistent between AD and SD
transformations.
Building on such consistency, I took the node ID to be the fragment name
(used to identify fragments and fragment references in AndreML) and
checked for previously created RDG nodes with the same ID (the class
`RDGNode` already kept track of that information).